MAJOR COMMIT - MOVEMENT
This is how it should function. Joystick should be a child of the camera so it stays static, but player should not be because player will interact with other things in the world, so instead just move the player and then update the camera position to player position so it appears to follow the player.